Get rid of RHModificationBaseProtected
It was only used for event management and thus does not belong into
rh/base.py
